ixgbe: Add a new netdev op to allow/prevent a VF from querying an RSS info
Implements the new netdev op to allow user to enable/disable the ability
of a specific VF to query its RSS Indirection Table and an RSS Hash Key.

This patch limits the new feature support to 82599 and x540 devices only.
Support for other devices will be added later.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>

New in v10:
   - United the for-each-VF loops in __ixgbe_enable_sriov() and
     ixgbe_configure_virtualization().
   - Use the EOPNOTSUPP error code instead of EPERM when operation is
     not supported.

New in v9:
   - Reduce the support to 82599 and x540 devices only.